requirements.txt
streamlit>=1.41.0
langchain-google-genai>=2.0.0
langchain>=0.3.0
langchain-community>=0.3.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
